Remove the '.tc' alternative name
This change is primarily driven by the unnecessary discussion
(bikeshedding) that the alternative short name caused. For example, see
[1].

Took this opportunity to improve the language quite a bit.

Removed the Examples section, and reference to it. The intent is to add
it back with actual examples, at some later point.

Also removed 'samples/' directory, and references to it,for the same
reason as above.

Provide more clarity and example of difference between a 'project' and a
'tool'.

[1]: https://news.ycombinator.com/item?id=36478361

Removed samples directory